refactor: move ai providers to windmill-ai
