+++ backend/windmill-ai/src/providers/anthropic.rs
@@ -1,7 +1,4 @@
-use async_trait::async_trait;
-use serde::{Deserialize, Serialize};
-use serde_json::value::RawValue;
-use windmill_ai::{
+use crate::{
     ai_google::parse_data_url,
     ai_providers::AIProvider,
     image_handler::prepare_messages_for_api,
@@ -10,6 +7,9 @@ use windmill_ai::{
     types::*,
     utils::{extract_text_content, should_use_structured_output_tool},
 };
+use async_trait::async_trait;
+use serde::{Deserialize, Serialize};
+use serde_json::value::RawValue;
 use windmill_common::{client::AuthedClient, error::Error};
 
 /// Anthropic API version for standard API
