+++ backend/windmill-ai/src/providers/other.rs
@@ -1,7 +1,4 @@
-use async_trait::async_trait;
-use serde::Serialize;
-use serde_json;
-use windmill_ai::{
+use crate::{
     ai_providers::AIProvider,
     image_handler::prepare_messages_for_api,
     query_builder::{BuildRequestArgs, ParsedResponse, QueryBuilder, StreamEventSink},
@@ -9,6 +6,9 @@ use windmill_ai::{
     types::*,
     utils::should_use_structured_output_tool,
 };
+use async_trait::async_trait;
+use serde::Serialize;
+use serde_json;
 use windmill_common::{client::AuthedClient, error::Error};
 
 #[derive(Serialize, Debug, Clone)]
