+++ backend/windmill-worker/src/ai/mod.rs
@@ -1,7 +1,6 @@
 // AI executor module structure
 // This module will contain all AI-related execution logic
 
-pub mod providers;
 pub mod query_builder;
 pub mod tools;
 pub mod utils;
