--- backend/windmill-worker/src/ai/providers/mod.rs
@@ -1,7 +1,0 @@
-pub mod anthropic;
-#[cfg(feature = "bedrock")]
-pub mod bedrock;
-pub mod google_ai;
-pub mod openai;
-pub mod openrouter;
-pub mod other;
